# openapi.yaml
openapi: 3.1.0
info:
  title: Granite Jev-compatible API
  version: 0.1.0
  description: |
    A local Jev-compatible structured evaluation API backed by Granite and
    llama.cpp. Questions are evaluated independently against the shared state.
    Probabilities are constrained softmax scores over the legal answer labels;
    they are not TypeSafe's proprietary calibrated probabilities.
servers:
  - url: http://127.0.0.1:8080
    description: Local server (default)
paths:
  /:
    post:
      operationId: evaluate
      summary: Evaluate structured questions against shared state
      requestBody:
        $ref: '#/components/requestBodies/EvaluateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EvaluateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/evaluate:
    post:
      operationId: evaluateV1
      summary: Evaluate structured questions against shared state
      requestBody:
        $ref: '#/components/requestBodies/EvaluateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EvaluateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/systemone:
    post:
      operationId: evaluateSystemOne
      summary: Evaluate structured questions against shared state
      requestBody:
        $ref: '#/components/requestBodies/EvaluateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EvaluateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /ai/run:
    post:
      operationId: cloudflareAiRun
      summary: Cloudflare AI-compatible evaluation route
      requestBody:
        $ref: '#/components/requestBodies/CloudflareEvaluateRequest'
      responses:
        '200':
          $ref: '#/components/responses/EvaluateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /health:
    get:
      operationId: health
      summary: Liveness check
      responses:
        '200':
          $ref: '#/components/responses/HealthResponse'
  /ready:
    get:
      operationId: ready
      summary: Readiness check
      responses:
        '200':
          $ref: '#/components/responses/HealthResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  requestBodies:
    EvaluateRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EvaluateInput'
    CloudflareEvaluateRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CloudflareEvaluateInput'
  responses:
    EvaluateResponse:
      description: Structured answers and token usage.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EvaluateResponse'
    BadRequest:
      description: Invalid JSON or malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnprocessableEntity:
      description: Request is valid JSON but fails semantic validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadGateway:
      description: The inference backend returned an error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceUnavailable:
      description: The inference worker is unavailable or queue is full.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    HealthResponse:
      description: Server is live and the model has loaded.
      content:
        text/plain:
          schema:
            type: string
            example: ok
  schemas:
    EvaluateInput:
      type: object
      required: [state, questions]
      properties:
        state:
          description: Shared state evaluated by every question.
          type: [string, number, boolean, object, array, 'null']
        questions:
          type: object
          minProperties: 1
          additionalProperties:
            $ref: '#/components/schemas/Question'
    CloudflareEvaluateInput:
      type: object
      required: [input]
      properties:
        model:
          type: string
          example: typesafe/jev
        input:
          $ref: '#/components/schemas/EvaluateInput'
    Question:
      discriminator:
        propertyName: type
        mapping:
          noul: '#/components/schemas/NoulQuestion'
          choice: '#/components/schemas/ChoiceQuestion'
          score: '#/components/schemas/ScoreQuestion'
      oneOf:
        - $ref: '#/components/schemas/NoulQuestion'
        - $ref: '#/components/schemas/ChoiceQuestion'
        - $ref: '#/components/schemas/ScoreQuestion'
    NoulQuestion:
      type: object
      required: [type, instructions]
      properties:
        type:
          const: noul
        instructions:
          type: string
          minLength: 1
        criteria:
          type: object
          required: ['true', 'false']
          minProperties: 2
          maxProperties: 2
          properties:
            'true':
              type: string
              nullable: true
            'false':
              type: string
              nullable: true
    ChoiceQuestion:
      type: object
      required: [type, instructions, criteria]
      properties:
        type:
          const: choice
        instructions:
          type: string
          minLength: 1
        criteria:
          type: object
          minProperties: 2
          maxProperties: 255
          additionalProperties:
            type: string
            nullable: true
    ScoreQuestion:
      type: object
      required: [type, instructions, criteria]
      properties:
        type:
          const: score
        instructions:
          type: string
          minLength: 1
        criteria:
          type: array
          minItems: 2
          maxItems: 10
          items:
            type: string
    EvaluateResponse:
      type: object
      required: [model, answers, usage]
      properties:
        model:
          type: string
          example: granite-jev-0.1.0
        answers:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Answer'
        usage:
          $ref: '#/components/schemas/Usage'
    Answer:
      discriminator:
        propertyName: type
        mapping:
          noul: '#/components/schemas/NoulAnswer'
          choice: '#/components/schemas/ChoiceAnswer'
          score: '#/components/schemas/ScoreAnswer'
      oneOf:
        - $ref: '#/components/schemas/NoulAnswer'
        - $ref: '#/components/schemas/ChoiceAnswer'
        - $ref: '#/components/schemas/ScoreAnswer'
    NoulAnswer:
      type: object
      required: [type, noul]
      properties:
        type:
          const: noul
        noul:
          type: number
          minimum: 0
          maximum: 1
    ChoiceAnswer:
      type: object
      required: [type, choice, confidence, probabilities]
      properties:
        type:
          const: choice
        choice:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        probabilities:
          type: object
          minProperties: 2
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
    ScoreAnswer:
      type: object
      required: [type, score, confidence, legend, probabilities]
      properties:
        type:
          const: score
        score:
          type: number
          minimum: 0
        confidence:
          type: number
          minimum: 0
          maximum: 1
        legend:
          type: object
          minProperties: 2
          additionalProperties:
            type: string
        probabilities:
          type: object
          minProperties: 2
          additionalProperties:
            type: number
            minimum: 0
            maximum: 1
    Usage:
      type: object
      required: [input_tokens, output_tokens]
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string